macOS port: native Apple Silicon build
Completes the unfinished non-Windows port so the client builds, runs and plays
natively on macOS. Most of these are latent bugs that had never run correctly on
a 64-bit little-endian target rather than macOS-specific issues; CrossUO reached
the same conclusions independently on several of them.

Crypto (each broke the game connection):
- Twofish compiled big-endian on a little-endian CPU. platform.h selected
  endianness with #ifdef _M_IX86, an MSVC-on-x86 macro clang never defines.
- typedef unsigned long u32 in aes.h is 8 bytes on LP64, doubling every buffer
  in the key schedule and smashing the stack in reKey().
- Blowfish p_box/s_box were unsigned long but copied into unsigned int tables:
  144 into 72 and 8192 into 4096 bytes, across 25 key tables.
- CTwofishCrypt::m_IP was an 8-byte long filled by a 4-byte memcpy.

Networking:
- select(h, ...) must be select(h + 1, ...); the client could send but never
  receive. Winsock ignores the argument, so this only broke off Windows.
- tcp_connect blocked the main thread with no timeout.
- SOCKADDR_IN aliased to in_addr, making sizeof() 4 instead of 16.
- Login seed passed a binary address to inet_aton, which expects a string.
- Relay to a private address now stays on the login host (NAT'd shards).

Protocol:
- ReadString returned NUL-padded fixed-width fields, so empty character slots
  looked occupied and logging in to one dropped the connection.
- Answer the Orion version query under 0xBF/0xFACE, repeated across the
  server's check window (Sphere only populates LOCAL.CHAR once m_pChar exists).
- Packet size table fix from upstream PR #97.

Runtime:
- CConfigManager's constructor created a GL texture before main() and before
  any GL context existed.
- The main loop only ran OnMainLoop() when an SDL event arrived, so networking
  advanced only while the mouse moved, and it span at 100% CPU when idle.
- SetTimer/KillTimer were no-ops; now backed by SDL_AddTimer.
- MidiInfoStruct packing left an 8-byte pointer unaligned, which the arm64
  linker refuses to link.
- WaveHeader used unsigned long, shifting every field after chunkSize.

Audio: every BASS_* call was a no-op macro, so the build was silent. Adds an
SDL_mixer backend leaving CSoundManager unchanged.

Input: all 27 OnTextInput handlers were unimplemented, SDL_StartTextInput was
never called, VK_* constants matched neither Win32 nor SDL (VK_BACK was space),
gump dragging always measured zero distance, and UTF-8 conversion returned
nothing so all cliloc text and outgoing chat were empty.

Window: SDL_WINDOWEVENT_* subtypes were compared against ev.type, so the whole
window-event block was dead code. Window is now resizable and the game view
tracks it.

Also: locale detection (the client defaulted to Russian), clipboard, URL
opening, command-line parsing (which returned nothing, so -login was ignored),
and a music stream leak on every track change.

Adds tools/ for building and running without the Windows launcher:
make_client_cuo.py generates the otherwise-launcher-only Client.cuo,
setup-macos.sh builds and wires up a UO install, version-sweep.sh tests
candidate Orion versions. See README-macos.md.

+++ OrionUO/Crypt/aes.h
@@ -78,7 +78,9 @@ INCLUDE IMPLEMENTATION SPECIFIC INFORMATION.
 */
 
 typedef unsigned char u8;
-typedef unsigned long u32; /* 32-bit unsigned quantity */
+// MUST be exactly 32 bits: "unsigned long" is 8 bytes on any LP64 target,
+// which doubled every buffer in the key schedule and overran the stack.
+typedef unsigned int u32; /* 32-bit unsigned quantity */
 typedef u32 fullSbox[4][256];
 
 /* The structure for key information */
